requirements.txt
psycopg2-binary
pandas
asyncpg
mysql-connector-python